% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{binomial_distribution_exactly}
\alias{binomial_distribution_exactly}
\title{Binomial probability}
\usage{
binomial_distribution_exactly(x, n, p)
}
\arguments{
\item{x}{number of successes we want}

\item{n}{number of times the experiment was done}

\item{p}{probability of success of the experiment}
}
\value{
probability that if we do the experiment \code{n} times we have \code{x} successes
}
\description{
Binomial probability
}
