% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{binomial_distribution_greater}
\alias{binomial_distribution_greater}
\title{Binomial probability X > x}
\usage{
binomial_distribution_greater(x, n, p, equal)
}
\arguments{
\item{x}{minimum number of successes we want}

\item{n}{number of times the experiment was done}

\item{p}{probability of success of the experiment}

\item{equal}{TRUE for including x (>=) FALSE to not include x (>)}
}
\value{
probability that if we do the experiment \code{n} times we have \code{x}or more successes
}
\description{
Binomial probability X > x
}
