% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{confidence_binomial}
\alias{confidence_binomial}
\title{Confidence interval for p, binomial distribution}
\usage{
confidence_binomial(n, phat, a)
}
\arguments{
\item{n}{size of the sample}

\item{phat}{probability of success in the sample}

\item{a}{confidence interval we want, for 95\% confidence, a = 0.05, for 90\% confidence, a = 0.1}
}
\value{
print confidence interval for the ACTUAL p in the distribution
}
\description{
Confidence interval for p, binomial distribution
}
