% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{confidence_diferenceMeans_var}
\alias{confidence_diferenceMeans_var}
\title{Confidence interval for difference of means, known population variances, normal distributions}
\usage{
confidence_diferenceMeans_var(mean1, mean2, var1, var2, n1, n2, a)
}
\arguments{
\item{mean1}{mean of the first sample}

\item{mean2}{mean of the second sample}

\item{var1}{variance of the first POPULATION (NOT SAMPLE)}

\item{var2}{variance of the second POPULATION (NOT SAMPLE)}

\item{n1}{size of the first sample}

\item{n2}{size of the second sample}

\item{a}{confidence interval we want, for 95\% confidence, a = 0.05, for 90\% confidence, a = 0.1}
}
\value{
print confidence interval for the difference of means in two normal distributions
}
\description{
Confidence interval for difference of means, known population variances, normal distributions
}
