% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{confidence_differenceProportions_binomial}
\alias{confidence_differenceProportions_binomial}
\title{Confidence interval for difference of (p1 - p2) of 2 binomial distributions}
\usage{
confidence_differenceProportions_binomial(p1, p2, n1, n2, a)
}
\arguments{
\item{p1}{proportion of success of the first sample}

\item{p2}{proportion of success of the second sample}

\item{n1}{size of the first sample}

\item{n2}{size of the second sample}

\item{a}{alpha of the confidence interval, for 95\% confidence, a = 0.05}
}
\description{
Confidence interval for difference of (p1 - p2) of 2 binomial distributions
}
