% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{confidence_normalMean_novar}
\alias{confidence_normalMean_novar}
\title{confidence interval for mean, unknown variance}
\usage{
confidence_normalMean_novar(n, mean, S, a)
}
\arguments{
\item{n}{size of the sample}

\item{mean}{mean of the sample}

\item{S}{standard deviation of the sample}

\item{a}{confidence interval we want, for 95\% confidence, a = 0.05, for 90\% confidence, a = 0.1}
}
\value{
print confidence interval for the ACTUAL mean in the distribution
}
\description{
confidence interval for mean, unknown variance
}
