% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{confidence_normalVariance}
\alias{confidence_normalVariance}
\title{confidence interval for variance, normal distribution}
\usage{
confidence_normalVariance(n, S, a)
}
\arguments{
\item{n}{size of the sample}

\item{S}{standard deviation of the sample}

\item{a}{confidence interval we want, for 95\% confidence, a = 0.05, for 90\% confidence, a = 0.1}
}
\value{
print confidence interval for the ACTUAL variance in the distribution
}
\description{
confidence interval for variance, normal distribution
}
