% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{confidence_ratioVariances_normal}
\alias{confidence_ratioVariances_normal}
\title{Confidence interval for difference of means, unknown variance}
\usage{
confidence_ratioVariances_normal(x, y, a)
}
\arguments{
\item{x}{dataset 1}

\item{y}{dataset 2}

\item{a}{confidence interval we want, for 95\% confidence, a = 0.05, for 90\% confidence, a = 0.1}
}
\value{
print confidence interval for the difference of means in two normal distributions
}
\description{
Confidence interval for difference of means, unknown variance
}
