% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{density_chisq}
\alias{density_chisq}
\title{Density of sample quasivariance (Marga)}
\usage{
density_chisq(x, n, type = c("density", "acumulative", "cuantil"))
}
\arguments{
\item{x}{is the value we want to knoe it's density}

\item{n}{are the grades of liberty for all sample quasivariance}

\item{type}{"density"=dquisq, "acumulative"=pshisq, "cuantil"= qchisq}
}
\description{
(Benja): no estoy muy seguro que hace esto la vd pero lo dejo por si ella la quiere para algo
}
