% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hypothesis_mean_normal}
\alias{hypothesis_mean_normal}
\title{Hypothesis test for mean of a normal distribution}
\usage{
hypothesis_mean_normal(
  x,
  mean0,
  type = c("two.sided", "less", "greater"),
  a,
  knownVar = NULL
)
}
\description{
Hypothesis test for mean of a normal distribution
}
