% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hypothesis_proportion}
\alias{hypothesis_proportion}
\title{Hypothesis test for a binomial proportion}
\usage{
hypothesis_proportion(
  x,
  n,
  p0,
  type = c("two.sided", "less", "greater"),
  a = 0.05
)
}
\arguments{
\item{x}{number of successes}

\item{n}{number of trials}

\item{p0}{hypothesized proportion}

\item{type}{is the alternative hypothesis: "two.sided" if H0 is = and H1 !=, "less" if H0 is > and H1 <, "greater" if H0 is < and H1 >}

\item{a}{is the confidence level, for confidence 95\%, a = 0.05}
}
\description{
Hypothesis test for a binomial proportion
}
