% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hypothesis_TwoMeans_normal}
\alias{hypothesis_TwoMeans_normal}
\title{Hypothesis test for mean of two samples}
\usage{
hypothesis_TwoMeans_normal(
  x,
  y,
  type = c("two.sided", "less", "greater"),
  a = 0.05,
  knownVar1 = NULL,
  knownVar2 = NULL
)
}
\arguments{
\item{x}{first dataset}

\item{y}{second dataset}

\item{type}{is the alternative hypothesis: "two.sided" if H0 is = and H1 !=, "less" if H0 is > and H1 <, "greater" if H0 is < and H1 >}

\item{a}{is the confidence level, for confidence 95\%, a = 0.05}

\item{knownVar1}{population variance of x (NULL if unknown)}

\item{knownVar2}{population variance of y (NULL if unknown)}
}
\value{
test output
}
\description{
Hypothesis test for mean of two samples
}
