% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hypothesis_TwoVariances}
\alias{hypothesis_TwoVariances}
\title{Hypothesis test for equality of variances (two normal pops)}
\usage{
hypothesis_TwoVariances(
  x,
  y,
  type = c("two.sided", "less", "greater"),
  a = 0.05
)
}
\arguments{
\item{x}{is the first dataset}

\item{y}{is the second dataset}

\item{type}{is the alternative hypothesis: "two.sided" if H0 is = and H1 !=, "less" if H0 is > and H1 <, "greater" if H0 is < and H1 >}

\item{a}{is the confidence level, for confidence 95\%, a = 0.05}
}
\description{
Hypothesis test for equality of variances (two normal pops)
}
