% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hypothesis_variance_normal}
\alias{hypothesis_variance_normal}
\title{Hypothesis test for variance of a normal distribution}
\usage{
hypothesis_variance_normal(
  x,
  var0,
  type = c("two.sided", "less", "greater"),
  a = 0.05
)
}
\arguments{
\item{x}{dataset}

\item{var0}{hypothesised variance}

\item{type}{is the alternative hypothesis: "two.sided" if H0 is = and H1 !=, "less" if H0 is > and H1 <, "greater" if H0 is < and H1 >}

\item{a}{is the confidence level, for confidence 95\%, a = 0.05}
}
\value{
prints test statistic and p-value (use p vs a to decide)
General rule: If p <= a  Reject H0, If p > a  Do NOT reject H0
}
\description{
Hypothesis test for variance of a normal distribution
}
