% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mean_sample_what_is_the_st_of_sample}
\alias{mean_sample_what_is_the_st_of_sample}
\title{This is for when we want the st of the sample mean (Marga)}
\usage{
mean_sample_what_is_the_st_of_sample(n, st, p)
}
\arguments{
\item{n}{size of the sample}

\item{st}{is the standart deviation of the population}

\item{is}{the size of the population}
}
\description{
This is for when we want the st of the sample mean (Marga)
}
