% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_proportion_big_n}
\alias{sample_proportion_big_n}
\title{Sample proportion if n>30 (Marga)}
\usage{
sample_proportion_big_n(x, n, p, lower)
}
\arguments{
\item{x}{the number of sucesses}

\item{n}{is the size of the sample(always over 30)}

\item{p}{proportional population}

\item{lower}{TRUE(lower than x) or FALSE(greater than x)}
}
\description{
Sample proportion if n>30 (Marga)
}
