% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_Quasivariance}
\alias{sample_Quasivariance}
\title{Sample Quasivariance(S^2) distribution}
\usage{
sample_Quasivariance(n, var, val, lower)
}
\arguments{
\item{n}{size of the sample}

\item{var}{variance of the population}

\item{val}{value that we want to be greater/lower than}

\item{lower}{TRUE(lower than val) or FALSE(greater than val)}
}
\description{
Sample Quasivariance(S^2) distribution
}
