% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sampleMean_finpop_var}
\alias{sampleMean_finpop_var}
\title{Sample mean with finite population, known variance}
\usage{
sampleMean_finpop_var(x, m, st, n, p)
}
\arguments{
\item{x}{is the value that we suposse the mean sample takes (or greater)}

\item{m}{is the mean of the population (we want to know the mean of the sample)}

\item{st}{is the standart deviation of the population}

\item{n}{size of the sample}

\item{p}{is the size of the population}
}
\description{
Sample mean with finite population, known variance
}
