% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sampleMean_ifpop_nbig}
\alias{sampleMean_ifpop_nbig}
\title{Sample mean with infinite population, unknown variance, n > 30}
\usage{
sampleMean_ifpop_nbig(m, var, n, val, lower)
}
\arguments{
\item{m}{media}

\item{n}{size of the sample.}

\item{val}{value that the mean has to be greater/lower than}

\item{lower}{TRUE(lower than val) or FALSE(greater than val)}

\item{S}{standard deviation of the sample}
}
\value{
probability that the mean of a sample is greater than \code{val}
}
\description{
Sample mean with infinite population, unknown variance, n > 30
}
\examples{
sampleMean_ifpop_nbig(6, 4, 25, 6.3, TRUE)
}
