% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sampleMean_ifpop_novar}
\alias{sampleMean_ifpop_novar}
\title{Sample mean with infinite population, unknown variance}
\usage{
sampleMean_ifpop_novar(m, S, n, val, lower)
}
\arguments{
\item{m}{mean of the population (NULL if you dont have)}

\item{S}{standard deviation of the sample}

\item{n}{size of the sample.}

\item{val}{value that the mean has to be greater/lower than}

\item{lower}{TRUE(lower than val) or FALSE(greater than val)}
}
\value{
probability that the mean of a sample is greater/lower than \code{val}
}
\description{
Sample mean with infinite population, unknown variance
}
\examples{
sampleMean_ifpop_novar(6, 4, 25, 6.3, TRUE)

#exercise example
sampleMean_ifpop_novar(12, 4, 25, 8, TRUE)
# probability that the mean of
#the sample is lower than 8 with std = 4 n = 25

#Calculate the probability that the mean of the sample does not differ from
#the population mean by more than 8 units, with std = 10.85 and n = 16

probMoreThanMinus8 <- sampleMean_ifpop_novar(NULL, 10.85, 16, -8, FALSE)
probMoreThan8 <- sampleMean_ifpop_novar(NULL, 10.85, 16, 8, FALSE)
probInBetween <- probMoreThanMinus8 - probMoreThan8
}
