% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sampleMean_ifpop_nsmall}
\alias{sampleMean_ifpop_nsmall}
\title{Sample mean with infinite population, unknown variance, n < 30}
\usage{
sampleMean_ifpop_nsmall(S, n, val, lower)
}
\arguments{
\item{S}{standard deviation of the sample}

\item{n}{size of the sample.}

\item{val}{value that the mean has to be greater/lower than}

\item{lower}{TRUE(lower than val) or FALSE(greater than val)}
}
\value{
probability that the mean of a sample is greater than \code{val}
}
\description{
Sample mean with infinite population, unknown variance, n < 30
}
\examples{
sampleMean_ifpop_nsmall(4, 25, 8, TRUE), probability that the mean of
the sample is lower than 8 with std = 4 n = 25

Calculate the probability that the mean of the sample does not differ from
the population mean by more than 8 units, with std = 10.85 and n = 16

probMoreThanMinus8 <- sampleMean_if_pop_nsmall(10.85, 16, -8, FALSE)
probMoreThan8 <- sampleMean_if_pop_nsmall(10.85, 16, 8, FALSE)
probInBetween <- probMoreThan8 - probMoreThanMinus8
}
