% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sampleMean_ifpop_var}
\alias{sampleMean_ifpop_var}
\title{Sample mean with infinite population, known variance}
\usage{
sampleMean_ifpop_var(m, var, n, val, lower)
}
\arguments{
\item{m}{media}

\item{var}{variance of the full population}

\item{n}{size of the sample.}

\item{val}{value that the mean has to be greater/lower than}

\item{lower}{TRUE(lower than val) or FALSE(greater than val)}
}
\value{
probability that the mean of a sample is greater than \code{val}
}
\description{
Sample mean with infinite population, known variance
}
\examples{
sampleMean_ifpop_var(6, 4, 25, 6.3, TRUE)
}
