% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{density_chisq}
\alias{density_chisq}
\title{Sample Quasivariance(S^2) distribution}
\usage{
density_chisq(x, n, type = c("density", "acumulative", "cuantil"))
}
\arguments{
\item{x}{is the value we want to knoe it's density}

\item{n}{are the grades of liberty for all sample quasivariance}

\item{type}{"density"=dquisq, "acumulative"=pshisq, "cuantil"= qchisq}
}
\description{
Sample Quasivariance(S^2) distribution
}
