% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_proportion_big_n}
\alias{sample_proportion_big_n}
\title{Sample proportion if n>30}
\usage{
sample_proportion_big_n(x, n, p, type = c("density", "cdf", "upper"))
}
\arguments{
\item{n}{is the size of the sample(always over 30)}

\item{p}{proporcional population}

\item{type}{"density" = dnorm, "cdf" = pnorm, "upper" = 1 - pnorm}

\item{is}{the number of sucesses}
}
\description{
Sample proportion if n>30
}
