% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sampleMean_finpop_novar}
\alias{sampleMean_finpop_novar}
\title{Sample mean with finite population, unknown variance and large n}
\usage{
sampleMean_finpop_novar(x, m, s, n, p)
}
\arguments{
\item{x}{is the value that we suposse the mean sample takes (or greater)}

\item{m}{is the mean of the population (we want to know the mean of the sample)}

\item{s}{is the desviacion muestral (ni idea de como era eso en inglés)}

\item{n}{size of the sample}

\item{p}{is the size of the population}
}
\description{
Sample mean with finite population, unknown variance and large n
}
